refactor into base class and subclass
