+++ scripts/screens/StartScreen.py
@@ -41,8 +41,8 @@ from ..housekeeping.datadir import get_cache_dir
 from ..housekeeping.update import has_update, UpdateChannel, get_latest_version_number
 from ..housekeeping.version import get_version_info
 from ..ui.generate_button import get_button_dict, ButtonStyles
-from ..ui.windows.tortie_patch_tool import TortiePatchToolWindow
-from ..ui.windows.white_patch_tool import WhitePatchToolWindow
+from scripts.ui.windows.dev_tool_windows.tortie_patch_tool import TortiePatchToolWindow
+from scripts.ui.windows.dev_tool_windows.white_patch_tool import WhitePatchToolWindow
 
 logger = logging.getLogger(__name__)
 has_checked_for_update = False
@@ -100,6 +100,8 @@ class StartScreen(Screens):
                 self.change_screen(GameScreen.EVENT_EDIT)
             elif element == self.elements.get("tortie_patch"):
                 TortiePatchToolWindow()
+            elif element == self.elements.get("white_patch"):
+                WhitePatchToolWindow()
             elif element == self.elements.get("bug_report"):
                 open_url("https://github.com/ClanGenOfficial/clangen/issues/new/choose")
             elif element == self.social_buttons["discord_button"]:
@@ -256,6 +258,17 @@ class StartScreen(Screens):
                 manager=MANAGER,
                 anchors={
                     "top_target": self.elements["event_edit"],
+                    "left_target": self.elements["continue"],
+                },
+            )
+            self.elements["white_patch"] = UISurfaceImageButton(
+                ui_scale(pygame.Rect((15, 15), (200, 30))),
+                "buttons.white_patch",
+                image_dict=get_button_dict(ButtonStyles.MAINMENU, (200, 30)),
+                object_id="@buttonstyles_mainmenu",
+                manager=MANAGER,
+                anchors={
+                    "top_target": self.elements["tortie_patch"],
                     "left_target": self.elements["continue"],
                 },
             )
@@ -449,5 +462,3 @@ class StartScreen(Screens):
 
         # LOAD settings
         game_settings_load()
-
-        WhitePatchToolWindow()
